Update and rename index.py to simple_search.py
